\name{modelling-package}
\alias{modelling-package}
\alias{modelling}
\docType{package}
\title{
\packageTitle{modelling}
}
\description{
\packageDescription{modelling}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{modelling}
\packageIndices{modelling}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{modelling}

Maintainer: \packageMaintainer{modelling}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
